requirements.txt
yt-dlp>=2024.3.10
Pillow>=10.3.0
groq>=0.5.0
click>=8.1.7
webvtt-py>=0.4.6
spacy>=3.7.0
scikit-learn>=1.4.0
